pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "gunicorn",
    "sqlalchemy",
    "python-dotenv",
    "psycopg2-binary"
]
